refactor: connectors/web split — webui/ owns the UI server, connectors/web/ keeps the chat connector
The 'connectors/web' folder had grown into the entire workbench HTTP
service (17 routes, ~2500 lines covering trading, market, topology,
diary, dev, persona, news, ...). Most of those routes have nothing to
do with chat — they exist because the React UI needs an API. Calling
them "connector" stretched the word past usefulness, and the directory
had become the catch-all for new HTTP surface.

Split:
- src/webui/                    NEW. Owns the Hono app, serves UI build,
                                hosts every HTTP route. Plugin name
                                renamed 'web' → 'webui'.
  ├── plugin.ts                 was connectors/web/web-plugin.ts
  ├── routes/                   was connectors/web/routes/
  └── __tests__/                was connectors/web/__tests__/
- src/connectors/web/           Slim. Now only the WebConnector class
                                — the actual chat-transport piece that
                                belongs alongside Telegram / MCP-Ask.

Mechanical move only — no business logic changes, no internal
reorganisation of routes. ~25 import paths rewritten to match the new
nesting depth (../../../core → ../../core for routes; ../../core →
../core for plugin.ts).

The chat routes (chat / channels / media) physically live in webui/
because that's where the HTTP host is, even though semantically they
implement the web chat connector's wire protocol. WebConnector class
stays in connectors/ as the contract object registered with
ConnectorCenter; webui/plugin.ts imports + registers it.

Future work (not in this PR): trading.ts and diary.ts hold real domain
logic (executeOneShot pipeline, FX-rate aggregation, heartbeat session
joining) that should move to domain/. Routes should be thin
HTTP-boundary wrappers over domain services.

1331 tests pass, tsc clean.
