+++ src/connectors/web/index.ts
@@ -1,3 +1,1 @@
-export { WebPlugin } from './web-plugin.js'
 export { WebConnector } from './web-connector.js'
-export type { WebConfig } from './web-plugin.js'
