+++ src/main.ts
@@ -5,7 +5,7 @@ import { loadConfig, readUTAsConfig } from './core/config.js'
 import type { Plugin, EngineContext, ReconnectResult } from './core/types.js'
 import { McpPlugin } from './server/mcp.js'
 import { TelegramPlugin } from './connectors/telegram/index.js'
-import { WebPlugin } from './connectors/web/index.js'
+import { WebPlugin } from './webui/index.js'
 import { McpAskPlugin } from './connectors/mcp-ask/index.js'
 import { createThinkingTools } from './tool/thinking.js'
 import { UTAManager, createSnapshotService, createSnapshotScheduler } from './domain/trading/index.js'
