+++ src/webui/index.ts
@@ -1,0 +1,2 @@
+export { WebPlugin } from './plugin.js'
+export type { WebConfig } from './plugin.js'
