+++ src/webui/routes/heartbeat.ts
@@ -1,7 +1,7 @@
 import { Hono } from 'hono'
 import { readFile, writeFile, mkdir } from 'node:fs/promises'
 import { dirname } from 'node:path'
-import type { EngineContext } from '../../../core/types.js'
+import type { EngineContext } from '../../core/types.js'
 
 const PROMPT_FILE = 'data/brain/heartbeat.md'
 const PROMPT_DEFAULT = 'default/heartbeat.default.md'
