+++ src/webui/routes/news.ts
@@ -1,5 +1,5 @@
 import { Hono } from 'hono'
-import type { EngineContext } from '../../../core/types.js'
+import type { EngineContext } from '../../core/types.js'
 
 const VALID_LOOKBACKS = new Set(['1h', '2h', '12h', '24h', '1d', '2d', '7d', '30d'])
 const DEFAULT_LOOKBACK = '24h'
