+++ src/webui/routes/trading-config.ts
@@ -1,11 +1,11 @@
 import { Hono } from 'hono'
-import type { EngineContext } from '../../../core/types.js'
+import type { EngineContext } from '../../core/types.js'
 import {
   readUTAsConfig, writeUTAsConfig,
   utaConfigSchema,
-} from '../../../core/config.js'
-import { createBroker } from '../../../domain/trading/brokers/factory.js'
-import { BUILTIN_BROKER_PRESETS } from '../../../domain/trading/brokers/presets.js'
+} from '../../core/config.js'
+import { createBroker } from '../../domain/trading/brokers/factory.js'
+import { BUILTIN_BROKER_PRESETS } from '../../domain/trading/brokers/presets.js'
 
 // ==================== Credential helpers ====================
